+++ CHANGELOG.md
@@ -12,12 +12,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added sliders to sidebar for precipitation duration and frequency
 - Added polygon layer to map
 - Polygon layer queried with slider
+- Application header
+- Dialog for About info
 
 ### Changed
 
 - Changed favicon to WIM favicon
+- Expansion panel styling
 
 ### Fixed
 
--
+- Removed vertical scrollbar